requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
psycopg2-binary==2.8.4
SQLAlchemy==1.3.11
Werkzeug==0.16.0
WTForms==2.2.1
